+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Add OutlierDampener weighting option
 
 ### v0.26.3
 * Add `Config.numParticipatedTagTootsPerTag` option and set it to 5 so we pull far less participated tag toots